+++ src/docs/tasks.md
@@ -14,10 +14,7 @@ What is a `Task`? See [`src/tasks/README.md`](../task).
 - [dev](../dev.task.ts) - start dev server
 - [format](../format.task.ts) - format source files
 - [gen](../gen.task.ts) - run code generation scripts
-- [project/build](../project/build.task.ts) - build, create, and link the distribution
-- [project/dist](../project/dist.task.ts) - create and link the distribution
 - [project/echo](../project/echo.task.ts) - diagnostic task that logs CLI args
-- [project/link](../project/link.task.ts) - link the distribution
 - [serve](../serve.task.ts) - start static file server
 - [server](../server.task.ts) - start API server
 - [start](../start.task.ts) - runs the dist/ builds for production
